Add HttpOnly and Secure flag to session cookie
HttpOnly cookie flag is added to the session cookie regardless of the
transport (HTTP / HTTPS).

Secure flag is added only if the service is exposed over HTTPS and if
the request used to create the session is made over the HTTS port.

Additionally, removed and unnecessary logic from TestHttpFutureListener
that re-add Cookie and CORS headers.

Improved tests to check HttpOnly flag.