Support local participants
Local participants in a transaction will be in the same process as the
initiator. Hence, there is no need to make calls over the network. So
the protocol endpoints become function pointers in such a case. Now when
there are nested transactions in the initiator, those transactions
become local participants.